lots of errors

+++ src/software/thunderscope/common/common_widgets.py
@@ -89,7 +89,7 @@ class ColorProgressBar(QProgressBar):
         self.decimals = 10 ** decimals
 
         super(ColorProgressBar, self).setRange(
-            min_val * self.decimals, max_val * self.decimals
+            int(min_val * self.decimals), int(max_val * self.decimals)
         )
 
         super(ColorProgressBar, self).setStyleSheet(
